Move `assert` and `record` functions about snapshot into a separate file
